@@ -1,2 +1,15 @@
-package org.example.algorithms;public class TwoDArrayTest {
+package org.example.algorithms;
+
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@DisplayName("Writing assertions for balance tests.")
+public class TwoDArrayTest {
+
+// add helper for reading from text file which contains test data.
+
+
 }
